pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelayer"
version = "0.1.0"
description = "Intelligent analytics layer for MSME ERPs (Tally-first) — HTTP/XML adapter, Postgres CDM, Metabase dashboards."
requires-python = ">=3.12"
dependencies = [
  "requests>=2.32",
  "lxml>=5.2",
  "pydantic>=2.7",
  "tenacity>=9.0",
  "loguru>=0.7",
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
  "jinja2>=3.1",
]

[tool.setuptools.packages.find]
include = ["adapters*", "agent*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true

